Conflicts:
	SConstruct
	bson/bsonelement.h
	bson/bsonobj.h
	bson/bsonobjbuilder.h
	bson/util/builder.h
	db/curop.h
	db/dbmessage.h
	db/diskloc.h
	db/geo/2d.cpp
	db/instance.cpp
	db/jsobj.cpp
	db/jsobj.h
	db/namespace.h
	db/oplog.cpp
	db/pdfile.h
	db/query.cpp
	db/recstore.h
	s/d_logic.cpp
	s/request.cpp
	tools/restore.cpp
	util/log.h
	util/message.cpp
	util/message.h
	util/optime.h

+++ db/tests.cpp
@@ -19,7 +19,7 @@
    unit test & such
 */
 
-#include "stdafx.h"
+#include "pch.h"
 #include "../util/mmap.h"
 
 namespace mongo {
@@ -32,7 +32,7 @@ namespace mongo {
 
         MemoryMappedFile f;
 
-        long len = 64*1024*1024;
+        unsigned long long len = 64*1024*1024;
         char *p = (char *) f.map("/tmp/test.dat", len);
         char *start = p;
         char *end = p + 64*1024*1024-2;
